test(stories): add breakpoint stories for the MyKiva carousels
Adds stories for LendingCategorySection and LendingStats, extends the
BorrowerCarousel set, and declares on each the three viewports where the slide
widths differ. BorrowerStatusCarousel.stories.js is renamed to match the
component it covers, and the JourneyCardCarousel slide fixtures move to a
shared mock so LendingStats can reuse them.

AD-439
