+++ .storybook/storybookStyles.scss
@@ -1,3 +1,19 @@
 body {
 	background: transparent;
 }
+
+.lending-stats-page-context {
+	.kv-carousel {
+		overflow: visible;
+	}
+
+	.kv-carousel > div:first-child {
+		column-gap: 16px;
+	}
+
+	@media (min-width: 64rem) {
+		.kv-carousel > div:first-child {
+			column-gap: 32px;
+		}
+	}
+}
